# OrigenX/BackEnd/openapi.yaml
openapi: 3.0.3
info:
  title: The Artisanal Connection API
  description: |
    API REST para The Artisanal Connection. Conecta productores de cafe colombiano con consumidores finales.

    ## Autenticacion
    La mayoria de endpoints requieren un Firebase ID Token en el header:
    `Authorization: Bearer <firebase_id_token>`

    ## Roles
    - **CONSUMER**: consumidor final que navega el catalogo y realiza pedidos.
    - **PRODUCER**: productor de cafe que gestiona productos y ventas.
    - **ADMIN**: administrador de la plataforma.

    ## Formato de errores
    Todos los errores siguen el formato uniforme:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Descripcion del error",
        "fields": [{"field": "email", "message": "El correo ya esta registrado"}]
      }
    }
    ```
  version: 0.1.0
  contact:
    name: The Artisanal Connection
servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo local
  - url: https://api.marketplacecafetero.com
    description: Servidor de produccion
...
security:
  - BearerAuth: []
paths:
...
  /api/cart:
    get:
      tags: [cart]
      summary: Obtener carrito
      description: Retorna el carrito del consumidor autenticado con subtotales y total calculados. Requiere rol CONSUMER. Requerimientos 12.1, 12.6.
      operationId: getCart
      responses:
        "200":
          description: Carrito del consumidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/cart/items:
    post:
      tags: [cart]
      summary: Agregar item al carrito
      description: Agrega un producto activo al carrito del consumidor. Retorna HTTP 400 si el producto esta inactivo. Requiere rol CONSUMER. Requerimientos 12.1, 8.4.
      operationId: addCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemCreate"
            example:
              product_id: prod123
              quantity: 2
      responses:
        "201":
          description: Item agregado al carrito
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Item agregado al carrito."
                  data:
                    $ref: "#/components/schemas/CartData"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/cart/items/{item_id}:
    put:
      tags: [cart]
      summary: Actualizar cantidad de item
      description: Actualiza la cantidad de un item en el carrito. Requiere rol CONSUMER. Requerimiento 12.2.
      operationId: updateCartItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
            example:
              quantity: 3
      responses:
        "200":
          description: Item actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Item actualizado."
                  data:
                    $ref: "#/components/schemas/CartData"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [cart]
      summary: Eliminar item del carrito
      description: Elimina un item del carrito y recalcula el total. Requiere rol CONSUMER. Requerimiento 12.3.
      operationId: deleteCartItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item eliminado del carrito
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Item eliminado del carrito."
                  data:
                    $ref: "#/components/schemas/CartData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID Token obtenido mediante Firebase Client SDK. Incluir en el header Authorization: Bearer <token>"
  parameters:
...
    ItemId:
      name: item_id
      in: path
      required: true
      description: Identificador unico del item del carrito
      schema:
        type: string
        example: item456mno
...
  schemas:
    ErrorField:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: El correo ya esta registrado
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Descripcion del error
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
    CartItemCreate:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          example: prod123abc
        quantity:
          type: integer
          minimum: 1
          example: 2
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          example: 3
    CartItemResponse:
      type: object
      properties:
        id:
          type: string
          example: item456mno
        product_id:
          type: string
          example: prod123abc
        product_name:
          type: string
          example: Cafe Especial Huila
        price:
          type: number
          example: 45000
        quantity:
          type: integer
          example: 2
        subtotal:
          type: number
          description: precio x cantidad
          example: 90000
        added_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
    CartData:
      type: object
      description: Datos del carrito del consumidor con subtotales calculados
      properties:
        user_id:
          type: string
          example: abc123uid
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemResponse"
        total:
          type: number
          description: Suma de todos los subtotales (precio x cantidad)
          example: 90000
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
    CartResponse:
      type: object
      description: Respuesta del carrito envuelta en data
      properties:
        data:
          $ref: "#/components/schemas/CartData"
...
  responses:
    ValidationError:
      description: Datos de entrada invalidos (HTTP 400)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Los datos de entrada son invalidos.
              fields:
                - field: email
                  message: El correo electronico no tiene un formato valido.
                - field: password
                  message: La contrasena debe tener al menos 8 caracteres.
    UnauthorizedError:
      description: No autenticado - JWT ausente o invalido (HTTP 401)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Token de autenticacion ausente o invalido.
    ForbiddenError:
      description: No autorizado - rol insuficiente o recurso ajeno (HTTP 403)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: No tienes permiso para realizar esta accion.
    NotFoundError:
      description: Recurso no encontrado (HTTP 404)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: El recurso solicitado no fue encontrado.
...
    InternalError:
      description: Error interno del servidor (HTTP 500)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: Error interno del servidor. Por favor intenta de nuevo.
...